Add task - Improved

+++ Controllers/TaskController.cs
@@ -41,25 +41,47 @@ namespace ProjectManagementApp.Controllers
 
             if (ModelState.IsValid)
             {
-                if (task.parentTask == null)
+                if (db.Projects.Any(x => x.name.ToLower() == task.projectName.ToLower()))
                 {
-                    task.parentTask = task.taskName;
+                    if (db.Users.Any(y => y.firstName.ToLower() == task.userName.ToLower()))
+                    {
+                        if (db.Tasks.Any(z => z.taskName.ToLower() == task.taskName.ToLower() && z.projectName.ToLower() == task.projectName.ToLower() && z.startDate == task.startDate && z.endDate == task.endDate))
+                        {
+                            ModelState.AddModelError("taskName", "Task Allready Present.");
+                        }
+                        else
+                        {
+                            if (task.parentTask == null)
+                            {
+                                task.parentTask = task.taskName;
+                            }
+
+
+                            var id = db.Projects.Where(x => x.name == task.projectName).Select(x => x.Id).FirstOrDefault();
+                            Project pro = db.Projects.Find(id);
+                            db.Projects.Remove(pro);
+                            db.SaveChanges();
+                            var taskn = Convert.ToInt32(pro.taskNo);
+                            var tasknu = taskn + 1;
+                            pro.taskNo = tasknu;
+                            db.Projects.Add(pro);
+                            db.SaveChanges();
+
+                            db.Tasks.Add(task);
+                            db.SaveChanges();
+                            return RedirectToAction("AddTask");
+                        }
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("userName", "User name not Present.");
+                    }
+                    
                 }
-
-
-                var id = db.Projects.Where(x => x.name == task.projectName).Select(x => x.Id).FirstOrDefault();
-                Project pro = db.Projects.Find(id);
-                db.Projects.Remove(pro);
-                db.SaveChanges();
-                var taskn = Convert.ToInt32(pro.taskNo);
-                var tasknu = taskn + 1;
-                pro.taskNo = tasknu;
-                db.Projects.Add(pro);
-                db.SaveChanges();
-
-                db.Tasks.Add(task);
-                db.SaveChanges();
-                return RedirectToAction("AddTask");
+                else
+                {
+                    ModelState.AddModelError("projectName", "Incorrect Project Name.");
+                }
             }
             var proj = db.Projects.Where(x => x.status == "In-Process").Select(r => r.name);
             var pt = db.Tasks.Select(r => r.parentTask).Distinct();
@@ -87,5 +109,29 @@ namespace ProjectManagementApp.Controllers
             return View(db.Tasks.Where(x => x.taskName.StartsWith(search) || search == null).ToList());
         }
 
+
+        //autocomplete project
+
+
+        [HttpPost]
+        public JsonResult GetSearchValue(string Prefix)
+        {
+            //Note : you can bind same list from database  
+            var name = (from c in db.Projects
+                             where (c.name.Contains(Prefix) && c.status.Equals("In-Process")) 
+                             select new { c.name}).ToList();
+            return Json(name, JsonRequestBehavior.AllowGet);
+        }
+
+        //autocomplete user
+
+        public JsonResult GetUser(string Prefix)
+        {
+            //Note : you can bind same list from database  
+            var firstName = (from c in db.Users
+                        where c.firstName.Contains(Prefix)
+                        select new { c.firstName }).ToList();
+            return Json(firstName, JsonRequestBehavior.AllowGet);
+        }
     }
 }
